enable using custom functon to map tags in method WithTags
This method allows using a function which can map a tag to its  original
value or to something else, such as its name. That way, when  using  the
method ::withTags, instead of attaching a Tag collection to each  model,
we can attach just the names of the tags by using a mapper function that
returns the tag name.